requirements.txt
# Python Version: 3.11.9
#
# Core UI and Framework
streamlit==1.54.0
streamlit-option-menu==0.4.0
extra-streamlit-components==0.1.81

# Data Manipulation and Visualization
pandas==2.3.3
plotly==6.5.2

# Database Connectivity (MySQL)
mysql-connector-python==9.6.0
SQLAlchemy==2.0.48

# Security
bcrypt==5.0.0

# Document Handling
python-docx==1.2.0
docx2pdf==0.1.8